+++ src/app/(user)/(account)/layout.tsx
@@ -1,4 +1,3 @@
-'use client';
 export default function MyPage({
   children,
 }: Readonly<{
