+++ src/app/(user)/my_page/page.tsx
@@ -1,6 +1,5 @@
-'use client';
 import ProductList from '@/components/products_list/products_list';
 
-export default function MyPage() {
+export default async function MyPage() {
   return <ProductList />;
 }
